Update all endpoints to use getCallerActions
