+++ stdlib/ballerina-builtin/src/main/ballerina/transactions/connector_participant_2pc.bal
@@ -43,7 +43,7 @@ public type Participant2pcClientEP object {
         self.conf = conf;
     }
 
-    public function getClient() returns Participant2pcClient {
+    public function getCallerActions() returns Participant2pcClient {
         Participant2pcClient client = new;
         client.clientEP = self;
         return client;
