+++ stdlib/ballerina-http/src/main/ballerina/http/non_listening_service_endpoint.bal
@@ -27,7 +27,7 @@ public type NonListener object{
     public native function initEndpoint () returns (error);
     public native function register (typedesc serviceType);
     public native function start ();
-    public native function getClient() returns Connection;
+    public native function getCallerActions() returns Connection;
     public native function stop ();
 };
