+++ stdlib/ballerina-http/src/main/ballerina/http/simple_client_endpoint.bal
@@ -51,7 +51,7 @@ public type SimpleClient object {
     documentation {
         Returns the backing HTTP client used by the endpoint.
     }
-    public function getClient() returns HttpClient {
+    public function getCallerActions() returns HttpClient {
         return httpEP.httpClient;
     }
