+++ stdlib/ballerina-http/src/main/ballerina/http/websocket_client.bal
@@ -31,7 +31,7 @@ public type WebSocketClient object {
 
     @Description {value:"Returns the connector that client code uses"}
     @Return {value:"The connector that client code uses"}
-    public function getClient () returns (WebSocketConnector)  {
+    public function getCallerActions () returns (WebSocketConnector)  {
         return self.conn;
     }
 
@@ -40,7 +40,7 @@ public type WebSocketClient object {
 
     @Description {value:"Stops the registered service"}
     public function stop () {
-        WebSocketConnector webSocketConnector = self.getClient();
+        WebSocketConnector webSocketConnector = self.getCallerActions();
         WebSocketConnectorError|() ignoredValue = webSocketConnector.close(1001, "The connection has been stopped");
     }
 };
