+++ stdlib/ballerina-jms/src/main/ballerina/mb/consumer.bal
@@ -24,7 +24,7 @@ public type ConsumerTemplate object {
 
     public function stop() {}
 
-    public function getClient() returns ConsumerConnector {
+    public function getCallerActions() returns ConsumerConnector {
         return new;
     }
