+++ stdlib/database/ballerina-h2/src/main/ballerina/h2/h2_endpoint.bal
@@ -65,7 +65,7 @@ public type Client object {
     documentation {
         Returns the connector that the client code uses.
     }
-    public function getClient() returns sql:CallerActions {
+    public function getCallerActions() returns sql:CallerActions {
         return self.h2Client;
     }
 };
