+++ tests/ballerina-test-integration/src/test/resources/grpc/client_streaming_client.bal
@@ -117,7 +117,7 @@ public type HelloWorldClient object {
         self.stub = stub;
     }
 
-    public function getClient () returns (HelloWorldStub) {
+    public function getCallerActions () returns (HelloWorldStub) {
         return self.stub;
     }
 };
