+++ tests/ballerina-test-integration/src/test/resources/grpc/grpc_secured_unary_client.bal
@@ -112,7 +112,7 @@ public type HelloWorldBlockingClient object {
         self.stub = stub;
     }
 
-    public function getClient() returns (HelloWorldBlockingStub) {
+    public function getCallerActions() returns (HelloWorldBlockingStub) {
         return self.stub;
     }
 };
@@ -134,7 +134,7 @@ public type HelloWorldClient object {
         self.stub = stub;
     }
 
-    public function getClient() returns (HelloWorldStub) {
+    public function getCallerActions() returns (HelloWorldStub) {
         return self.stub;
     }
 };
