+++ tests/ballerina-test-integration/src/test/resources/grpc/server_streaming_client.bal
@@ -103,7 +103,7 @@ public type HelloWorldClient object {
         self.stub = stub;
     }
 
-    public function getClient () returns (HelloWorldStub) {
+    public function getCallerActions () returns (HelloWorldStub) {
         return self.stub;
     }
 };
