+++ tests/ballerina-test-integration/src/test/resources/grpc/unary1_blocking_client.bal
@@ -277,7 +277,7 @@ public type HelloWorldBlockingClient object {
         self.stub = stub;
     }
 
-    public function getClient () returns (HelloWorldBlockingStub) {
+    public function getCallerActions () returns (HelloWorldBlockingStub) {
         return self.stub;
     }
 };
@@ -300,7 +300,7 @@ public type helloWorldClient object {
         self.stub = stub;
     }
 
-    public function getClient () returns (HelloWorldStub) {
+    public function getCallerActions () returns (HelloWorldStub) {
         return self.stub;
     }
 };
