+++ tests/ballerina-test-integration/src/test/resources/grpc/unary1_nonblocking_client.bal
@@ -246,7 +246,7 @@ public type HelloWorldBlockingClient object {
         self.stub = stub;
     }
 
-    public function getClient () returns (HelloWorldBlockingStub) {
+    public function getCallerActions () returns (HelloWorldBlockingStub) {
         return self.stub;
     }
 };
@@ -269,7 +269,7 @@ public type HelloWorldClient object {
         self.stub = stub;
     }
 
-    public function getClient () returns (HelloWorldStub) {
+    public function getCallerActions () returns (HelloWorldStub) {
         return self.stub;
     }
 };
