+++ tests/ballerina-test/src/test/resources/test-src/endpoint/testEndpointInFunction.bal
@@ -32,7 +32,7 @@ type DummyEndpoint object {
         exFlow = exFlow + "register:DummyEndpoint;";
     }
 
-    public function getClient () returns (DummyClient) {
+    public function getCallerActions () returns (DummyClient) {
         exFlow = exFlow + "getClient:DummyEndpoint;";
         return new;
     }
