+++ tests/ballerina-test/src/test/resources/test-src/endpoint/testEndpointWithService.bal
@@ -24,7 +24,7 @@ type DummyEndpoint object {
         exFlow = exFlow + "register:DummyEndpoint;";
     }
 
-    public function getClient () returns (DummyClient) {
+    public function getCallerActions () returns (DummyClient) {
         exFlow = exFlow + "getClient:DummyEndpoint;";
         return new;
     }
