+++ tests/ballerina-test/src/test/resources/test-src/net/http/resiliency/circuit-breaker-test.bal
@@ -46,7 +46,7 @@ function testTypicalScenario () returns (http:Response[] , http:HttpConnectorErr
     http:Response[] responses = [];
     http:HttpConnectorError[] errs = [];
     int counter = 0;
-    http:CircuitBreakerClient cbClient = check <http:CircuitBreakerClient>backendClientEP.getClient();
+    http:CircuitBreakerClient cbClient = check <http:CircuitBreakerClient>backendClientEP.getCallerActions();
     MockClient mockClient = new;
     cbClient.httpClient = <http:HttpClient> mockClient;
 
@@ -94,7 +94,7 @@ function testTrialRunFailure () returns (http:Response[] , http:HttpConnectorErr
     http:Response[] responses = [];
     http:HttpConnectorError[] errs = [];
     int counter = 0;
-    http:CircuitBreakerClient cbClient = check <http:CircuitBreakerClient>backendClientEP.getClient();
+    http:CircuitBreakerClient cbClient = check <http:CircuitBreakerClient>backendClientEP.getCallerActions();
     MockClient mockClient = new;
     cbClient.httpClient = <http:HttpClient> mockClient;
 
@@ -142,7 +142,7 @@ function testHttpStatusCodeFailure () returns (http:Response[] , http:HttpConnec
     http:Response[] responses = [];
     http:HttpConnectorError[] errs = [];
     int counter = 0;
-    http:CircuitBreakerClient cbClient = check <http:CircuitBreakerClient>backendClientEP.getClient();
+    http:CircuitBreakerClient cbClient = check <http:CircuitBreakerClient>backendClientEP.getCallerActions();
     MockClient mockClient = new;
     cbClient.httpClient = <http:HttpClient> mockClient;
